gettimeofday: Use uint64_t for timevals

+++ src/tasks/syscalls/gettimeofday.c
@@ -21,8 +21,8 @@
 #include <lib/time.h>
 
 struct timeval {
-	uint32_t tv_sec;
-	uint32_t tv_usec;
+	uint64_t tv_sec;
+	uint64_t tv_usec;
 };
 
 SYSCALL_HANDLER(gettimeofday) {
